requirements.txt
anthropic>=0.25.0
boto3>=1.34.0
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.13.0
requests>=2.31.0
python-dotenv>=1.0.0
